Move AssertEqualsOrAssertSameFloatParameterToSpecificMethodsTypeRector to CodeQuality
